Reorganize files and rename santec directory
