+++ python_il_sts/daq_device_class.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 DAQ Device Class.
 
@@ -14,7 +12,6 @@ from .error_handling_class import InstrumentError, instrument_error_strings
 
 # Import program logger
 from .logger import get_logger
-
 logger = get_logger("Spu Device Class.")
