+++ python_il_sts/error_handling_class.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 Error Handling Class.
