+++ python_il_sts/file_saving.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 File Logging.
 Save STS parameter data.
