+++ python_il_sts/mpm_instrument_class.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 MPM Instrument Class.
 
@@ -16,7 +14,6 @@ from .get_address import Instrument
 
 # Import program logger
 from .logger import get_logger
-
 logger = get_logger("MPM Instrument Class.")
